\name{fix_common_encoding_problems}
\alias{fix_common_encoding_problems}
\title{Fix common encoding problems when working with web imported data.}
\usage{fix_common_encoding_problems(x)}
\description{Fix common encoding problems when working with web imported data.}
\details{This function tries to resolve typical encoding problems when importing web data on Windows.
Typical problems occur with pound and hyphen (-), especially when these originated in MS-Word.}
\keyword{encoding}
\arguments{\item{x}{A character vector}}
