\name{fix_levels_01_SPSS}
\alias{fix_levels_01_SPSS}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{fix_levels_01_SPSS(dat)}
\description{Fix level formatting of all question with Yes/No type answers.}
\keyword{"clean}
\keyword{data"}
\arguments{\item{dat}{Surveyor data object}}
