\name{leveltest_R}
\alias{leveltest_R}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{leveltest_R(x)}
\description{Fix level formatting of all question with Yes/No type answers.}
\keyword{"clean}
\keyword{data"}
\seealso{\code{\link{fix_levels_01_SPSS}}, \code{\link{fix_levels_01_R}}}
\arguments{\item{x}{Surveyor data object}}
