\name{surveyor_extract}
\alias{surveyor_extract}
\title{Extract block of questions from surveyor object.}
\usage{surveyor_extract(surveyor, extract)}
\description{Extract block of questions from surveyor object.}
\keyword{tools}
\arguments{\item{surveyor}{Surveyor object}
\item{extract}{Character vector of question names to extract}}
