\name{surveyor_merge}
\alias{surveyor_merge}
\title{Merge surveyor data objects.}
\usage{surveyor_merge(dat1, dat2, ...)}
\description{Merge surveyor data objects.}
\details{The base R merge will merge data but not all of the attributes.  This function also merges the variable.labels attribute.}
\keyword{tools}
\arguments{\item{dat1}{Data frame}
\item{dat2}{Data frame}
\item{...}{Other parameters passed to merge}}
