\name{surveyor_rename_cols}
\alias{surveyor_rename_cols}
\title{Renames surveyor data columns.}
\usage{surveyor_rename_cols(dat, colnames, newnames)}
\description{Renames surveyor data columns.}
\details{When renaming surveyor data objects, it's important to also rename the variable.labels attribute.}
\keyword{tools}
\arguments{\item{dat}{Surveyor data object}
\item{colnames}{Character vector with column names to modify}
\item{newnames}{Character vector with replacement names}}
