\name{leveltest_SPSS}
\alias{leveltest_SPSS}
\title{Fix level formatting of all question with Yes/No type answers.}
\usage{
  leveltest_SPSS(x)
}
\arguments{
  \item{x}{Surveyor data object}
}
\description{
  Fix level formatting of all question with Yes/No type
  answers.
}
\seealso{
  \code{\link{fix_levels_01_SPSS}},
  \code{\link{fix_levels_01_R}}
}
\keyword{"clean}
\keyword{data"}

